online_inference/requirements.txt
pandas==1.2.4
uvicorn==0.13.4
requests==2.25.1
pydantic==1.6.1
pytest==6.2.3
joblib==0.17.0
numpy==1.20.1
hydra_core==1.0.6
omegaconf==2.0.6
dataclasses==0.8
fastapi==0.63
scikit_learn==0.24.2
